Install pip with get-pip.py instead of apt-get
We failed to build the docker since the pip version may be too old when installed with apt-get (e.g., 8.1, in ubuntu 16.04).

To fix this issue, we modify the ``setup.sh`` where we install pip with ``https://bootstrap.pypa.io/get-pip.py`` instead of ``apt-get install python-pip``. This works well for me.